+++ src/org/labkey/test/tests/core/login/DbReauthTest.java
@@ -2,14 +2,13 @@ package org.labkey.test.tests.core.login;
 
 import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
-import org.labkey.test.Locator;
 import org.labkey.test.categories.Daily;
-import org.labkey.test.tests.AbstractReauthTest;
+import org.labkey.test.tests.AbstractLoginFormReauthTest;
 import org.labkey.test.util.PasswordUtil;
 import org.labkey.test.util.TestUser;
 
 @Category({Daily.class})
-public class DbReauthTest extends AbstractReauthTest
+public class DbReauthTest extends AbstractLoginFormReauthTest
 {
     private static final TestUser USER1 = new TestUser("db_user1@reauth.test");
     private static final TestUser USER2 = new TestUser("db_user2@reauth.test");
@@ -17,24 +16,6 @@ public class DbReauthTest extends AbstractReauthTest
     public DbReauthTest()
     {
         super(new User(USER1.getEmail(), PasswordUtil.getPassword()), new User(USER2.getEmail(), PasswordUtil.getPassword()));
-    }
-
-    @Override
-    protected void clickSignIn()
-    {
-        clickAndWait(Locator.tagWithClass("a", "header-link").withText("Sign In"));
-    }
-
-    @Override
-    protected void authenticate(String email, String password)
-    {
-        doAndWaitForPageToLoad(() -> fillSignInFormAndSubmit(null, email, password));
-    }
-
-    @Override
-    protected void authenticateExpectingError(String email, String password)
-    {
-        fillSignInFormAndSubmit(null, email, password);
     }
 
     @Override
@@ -53,7 +34,6 @@ public class DbReauthTest extends AbstractReauthTest
     public static void setupProject() throws Exception
     {
         DbReauthTest init = getCurrentTest();
-
         init.doSetup();
     }
 
@@ -62,5 +42,4 @@ public class DbReauthTest extends AbstractReauthTest
         USER1.create(this).setInitialPassword();
         USER2.create(this).setInitialPassword();
     }
-
 }
